Fixing delete post

+++ sources/backend/repositories/postsRepository.py
@@ -58,16 +58,16 @@ class PostsRepository:
         finally:
             connection.close()
 
-    def delete_post(self, input_post):
-        print(input_post)
+    def delete_post(self, username, post_id):
         connection = self.__create_connection()
-        post_id = input_post["post_id"]
         if self.is_post_already_exists(post_id) is True:
             try:
                 cursor = connection.cursor()
-                request = f"DELETE FROM Posts WHERE id='{post_id}';"
-                cursor.execute(request)
-                return "Delete post successful", 200
+                request = f"DELETE FROM Posts WHERE id='{post_id}' AND author='{username}';"
+                success = cursor.execute(request) != 0
+                if success:
+                    return "Delete post successful", 200
+                return "Post does not belong to requesting user", 400
             finally:
                 connection.close()
         else:
